Re-instate the overloading removed by 8d15edf49
That was to make numeric comparison work on DBM::Deep objects. The
overloading code did not appear to be doing anything, so I removed it.
It turns out that it was working around a bug in perl 5.8.8 and ear-
lier with overload: blessing a reference into a class that has over-
load (like DBM::Deep::Null) does not active overloading if the refer-
ence did not already have it. So this adds it back to DBM::Deep, while
fixing it to make numeric comparison work properly (simply by omitting
the '""' entry).