Fix fieldhash usage to avoid a memory leak.
I forgot how to use field hashes. What was I thinking? (Or maybe I
should leave out the ‘what’!)

A field hash entry won’t be freed when the object is destroyed, unless
the object is registered. It won’t get registered unless the object
itself (as opposed to its address/id) is used as the key.