Put in a proper check for too many transactions.
Up till now, DBM::Deep has allowed more transactions to be started
than the file was created to hold. It would just start corrupting the
file, as the number of transactions determines the size of a bucket
list. See bucket_size in DBM::Deep::Sector::File::BucketList.

This also corrects one of the tests, which was starting one transac-
tion too many.