[rt.cpan.org #77746] Circular refs due to caching
A hash or array object return from a database is stored in the data-
base engine’s cache and reused the next time the same aggregate is
fetch, so that comparisons like $db->{h} == $db->{h} will work.

This cached item, like all database objects, holds a reference to the
database engine, which references its cache.  So we end up with a cir-
cular reference.

This commit makes the cache hold weak references.  It was already
doing that for external_refs mode, but it turns out we always need it.

This does mean that the same hash retrieved multiple times within a
single session will appear to change memory address, but no sane code
should be relying on that.  For it to break it would have to store a
stringified (or nummified) aggregate, while letting the actual aggre-
gate retrieved from the database be freed.

I also added an extra test to make sure stale cache entries are not
mistakenly reused, causing $db->{h} to return undef instead of what-
ever is stored there.